requirements.txt
python-nmap
requests
shodan
python-whois
